Controllers, Services, Repos y DTOs

+++ E-Commerce/src/main/java/com/uade/tpo/E_Commerce/entity/Category.java
@@ -15,7 +15,7 @@ public class Category {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id_category;
+    private Long id_category;
 
     @Column(name = "name_category", length = 60)
     private String name_category;
